requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
datetime-truncate==1.1.1
Django==4.1.4
django-admin-charts==1.2.4
django-bower==5.2.0
django-cors-headers==3.13.0
django-countries==7.5
django-filter==22.1
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
importlib-metadata==5.2.0
Jinja2==3.1.2
mariadb==1.1.5.post3
Markdown==3.4.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==22.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
Unidecode==1.3.6
zipp==3.11.0
